requirements.txt
customtkinter
Pillow
httpx
python-dotenv
